remove unneeded methods from SwapChain

+++ application/source/application_lod.cpp
@@ -434,7 +434,7 @@ void ApplicationLod::updateDescriptors() {
 
 void ApplicationLod::createDescriptorPools() {
   DescriptorPoolInfo info_pool{};
-  info_pool.reserve(m_shaders.at("lod"), 0, m_frame_resources.size());
+  info_pool.reserve(m_shaders.at("lod"), 0, uint32_t(m_frame_resources.size()));
   info_pool.reserve(m_shaders.at("lod"), 1, 1);
 
   m_descriptor_pool = DescriptorPool{m_device, info_pool};
